Move saveTexture function to utils
